Refactor Ownership and OwnerState
* Moved these two structs to internal
  folder.
