+++ hadoop-hdfs-project/hadoop-hdfs-native-client/src/main/native/libhdfspp/tools/hdfs-chown/hdfs-chown.h
@@ -23,8 +23,8 @@
 
 #include <boost/program_options.hpp>
 
-#include "hdfs-ownership.h"
 #include "hdfs-tool.h"
+#include "internal/hdfs-ownership.h"
 
 namespace hdfs::tools {
 /**
